move shared code to iris_data.py
